start refactoring code to work as a python package, added tests and updated the requirements. Still need to test if everything works as intended
